# Use serde_json serializer