`serde_json` serialization is faster than `ciborium`, so let's use this one instead

Before:
```
------------ 1 thread(s) -------------
**read_heavy** with prefill_fraction 0.95
ValueStorage:
1572864 operations across 1 thread(s) in 903.186041ms; time/op = 574ns

**insert_heavy** with prefill_fraction 0.0
ValueStorage:
1572864 operations across 1 thread(s) in 1.950499667s; time/op = 1.24µs

**update_heavy** with prefill_fraction 0.5
ValueStorage:
1572864 operations across 1 thread(s) in 1.766226833s; time/op = 1.122µs
```


After:
```
------------ 1 thread(s) -------------
**read_heavy** with prefill_fraction 0.95
ValueStorage:
1572864 operations across 1 thread(s) in 767.020958ms; time/op = 487ns

**insert_heavy** with prefill_fraction 0.0
ValueStorage:
1572864 operations across 1 thread(s) in 1.811419625s; time/op = 1.151µs

**update_heavy** with prefill_fraction 0.5
ValueStorage:
1572864 operations across 1 thread(s) in 1.68138975s; time/op = 1.068µs
```